% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_to_excel.R
\name{expr_to_excel}
\alias{expr_to_excel}
\title{Translate an R expression into an Excel conditional formatting formula}
\usage{
expr_to_excel(x, data, row_start = 2L)
}
\arguments{
\item{x}{An R-style expression}

\item{data}{A data frame that expression \code{x} relates to}

\item{row_start}{Integer reflecting the first row of data in the Excel sheet
to be output. Defaults to 2.}
}
\value{
A character string reflecting an Excel conditional formatting formula
}
\description{
Translate an R expression into an Excel conditional formatting formula
}
\examples{
library(datasets)

expr_to_excel(cyl > 4, mtcars)
expr_to_excel(cyl > 4 & hp < 200, mtcars)

}
