% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect.R
\name{qprotect}
\alias{qprotect}
\title{Protect one ore more columns of a worksheet}
\usage{
qprotect(
  password = NULL,
  cols = everything(),
  protect = TRUE,
  lockSelectingLockedCells = FALSE,
  lockSelectingUnlockedCells = FALSE,
  lockFormattingCells = FALSE,
  lockFormattingColumns = FALSE,
  lockFormattingRows = FALSE,
  lockInsertingColumns = TRUE,
  lockInsertingRows = TRUE,
  lockInsertingHyperlinks = FALSE,
  lockDeletingColumns = TRUE,
  lockDeletingRows = TRUE,
  lockSorting = FALSE,
  lockAutoFilter = FALSE,
  lockPivotTables = NULL,
  lockObjects = NULL,
  lockScenarios = NULL
)
}
\arguments{
\item{password}{(optional) password required to unprotect the worksheet}

\item{cols}{Tidy-selection specifying the columns that protection should
apply to. Defaults to \code{\link[dplyr:reexports]{dplyr::everything}} to select all columns.}

\item{protect}{Whether to protect or unprotect the sheet (default=TRUE)}

\item{lockSelectingLockedCells}{Whether selecting locked cells is locked}

\item{lockSelectingUnlockedCells}{Whether selecting unlocked cells is locked}

\item{lockFormattingCells}{Whether formatting cells is locked}

\item{lockFormattingColumns}{Whether formatting columns is locked}

\item{lockFormattingRows}{Whether formatting rows is locked}

\item{lockInsertingColumns}{Whether inserting columns is locked}

\item{lockInsertingRows}{Whether inserting rows is locked}

\item{lockInsertingHyperlinks}{Whether inserting hyperlinks is locked}

\item{lockDeletingColumns}{Whether deleting columns is locked}

\item{lockDeletingRows}{Whether deleting rows is locked}

\item{lockSorting}{Whether sorting is locked}

\item{lockAutoFilter}{Whether auto-filter is locked}

\item{lockPivotTables}{Whether pivot tables are locked}

\item{lockObjects}{Whether objects are locked}

\item{lockScenarios}{Whether scenarios are locked}
}
\description{
Wrapper to \code{\link[openxlsx:protectWorksheet]{openxlsx::protectWorksheet}} with additional argument \code{cols} to
enable protection to be limited to specific columns.
}
