% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwrite.R
\name{qwrite}
\alias{qwrite}
\title{Write workbook to an xlsx file}
\usage{
qwrite(wb, file, overwrite = FALSE)
}
\arguments{
\item{wb}{A Workbook object to write to file}

\item{file}{A character string naming an xlsx file}

\item{overwrite}{If \code{TRUE}, overwrite any existing file.}
}
\description{
Wrapper to \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook}} to write an Excel workbook to file
}
