% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect.R
\name{qprotect}
\alias{qprotect}
\title{Protect one ore more columns of a worksheet}
\usage{
qprotect(
  password = NULL,
  cols = everything(),
  row_buffer = 0L,
  protect = TRUE,
  lockSelectingLockedCells = FALSE,
  lockSelectingUnlockedCells = FALSE,
  lockFormattingCells = FALSE,
  lockFormattingColumns = FALSE,
  lockFormattingRows = FALSE,
  lockInsertingColumns = TRUE,
  lockInsertingRows = TRUE,
  lockInsertingHyperlinks = FALSE,
  lockDeletingColumns = TRUE,
  lockDeletingRows = TRUE,
  lockSorting = FALSE,
  lockAutoFilter = FALSE,
  lockPivotTables = NULL,
  lockObjects = NULL,
  lockScenarios = NULL
)
}
\arguments{
\item{password}{(optional) password required to unprotect the worksheet}

\item{cols}{Tidy-selection specifying the columns that protection should
apply to. Defaults to \code{\link[dplyr:reexports]{dplyr::everything}} to select all columns.}

\item{row_buffer}{The number of additional rows (beyond the range of the
current data) to unprotect within columns not specified in argument \code{cols}.
See explanation in Description. Defaults to \code{0}.}

\item{protect}{Whether to protect or unprotect the sheet (default=TRUE)}

\item{lockSelectingLockedCells}{Whether selecting locked cells is locked}

\item{lockSelectingUnlockedCells}{Whether selecting unlocked cells is locked}

\item{lockFormattingCells}{Whether formatting cells is locked}

\item{lockFormattingColumns}{Whether formatting columns is locked}

\item{lockFormattingRows}{Whether formatting rows is locked}

\item{lockInsertingColumns}{Whether inserting columns is locked}

\item{lockInsertingRows}{Whether inserting rows is locked}

\item{lockInsertingHyperlinks}{Whether inserting hyperlinks is locked}

\item{lockDeletingColumns}{Whether deleting columns is locked}

\item{lockDeletingRows}{Whether deleting rows is locked}

\item{lockSorting}{Whether sorting is locked}

\item{lockAutoFilter}{Whether auto-filter is locked}

\item{lockPivotTables}{Whether pivot tables are locked}

\item{lockObjects}{Whether objects are locked}

\item{lockScenarios}{Whether scenarios are locked}
}
\description{
Wrapper to \code{\link[openxlsx:protectWorksheet]{openxlsx::protectWorksheet}} with additional argument \code{cols} to
enable protection to be limited to specific columns.

In practice, protection is first applied to the entire worksheet, and then
subsequently columns not selected for protection (if any) are unlocked one by
one. This unlock step (when relevant) also requires a row specification,
which by default we limit to the range of the current data. Thus, in
'unprotected' columns within protected worksheets, rows beyond the range of
the data will remain protected. As a hack to work around this, the user can
specify a 'buffer' of additional empty rows to unprotect within each
non-protected column (e.g. to allow further data entry).
}
